Move format_minutes to chia.cmds.cmds_util
