+++ chia/cmds/farm_funcs.py
@@ -5,7 +5,7 @@ import traceback
 from pathlib import Path
 from typing import Any, Dict, List, Optional
 
-from chia.cmds.cmds_util import format_bytes, get_any_service_client
+from chia.cmds.cmds_util import format_bytes, format_minutes, get_any_service_client
 from chia.cmds.units import units
 from chia.consensus.block_record import BlockRecord
 from chia.rpc.farmer_rpc_client import FarmerRpcClient
@@ -13,7 +13,6 @@ from chia.rpc.full_node_rpc_client import FullNodeRpcClient
 from chia.rpc.wallet_rpc_client import WalletRpcClient
 from chia.util.default_root import DEFAULT_ROOT_PATH
 from chia.util.errors import CliRpcConnectionError
-from chia.util.misc import format_minutes
 from chia.util.network import is_localhost
 
 SECONDS_PER_BLOCK = (24 * 3600) / 4608
